# packages/api-types/schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "nof1-causal-lab episode API",
...
    "version": "0.1.0"
  },
  "paths": {
    "/api/episodes/{workspace_id}": {
      "get": {
        "summary": "Get Episode",
        "description": "Current episode state: the single read to poll while navigating.\n\nReturns per-artifact freshness (existence, staleness, version, provenance),\nthe `legal` moves available right now, and `auto_running` \u2014 whether the\nbackground driver is active. Replayed from the append-only transition log,\nso it works even against a published read-only store.",
        "operationId": "get_episode_api_episodes__workspace_id__get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Episode Api Episodes  Workspace Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/episodes/{workspace_id}/timeline": {
      "get": {
        "summary": "Get Timeline",
        "description": "The transition journal: every move attempt in order.\n\nEach record is `applied` (state advanced), `rejected` (illegal move, state\nunchanged), or `raised` (the transition ran but threw \u2014 the record carries the\ntyped error). Re-running after a `raised`/`rejected` is just proposing the\nmove again.",
        "operationId": "get_timeline_api_episodes__workspace_id__timeline_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Timeline Api Episodes  Workspace Id  Timeline Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/episodes/{workspace_id}/events": {
      "get": {
        "summary": "Get Events",
        "description": "Fine-grained telemetry (e.g. extraction worker fan-out, transition progress).\n\nPass the last-seen event id as `after` to page forward; omit it for the full\nstream. This is finer-grained than the timeline, which records only whole\nmove outcomes.",
        "operationId": "get_events_api_episodes__workspace_id__events_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Events Api Episodes  Workspace Id  Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/episodes/{workspace_id}/artifacts/{artifact_id}": {
      "get": {
        "summary": "Get Artifact",
        "description": "One artifact version: meta + inline JSON payloads.\n\nDefaults to the episode's *current* version from replayed applied transition\neffects. Binary payload files (parquet, pickle) are listed by name, never\ninlined.",
        "operationId": "get_artifact_api_episodes__workspace_id__artifacts__artifact_id__get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "question",
                "raw_data",
                "latent_structure",
                "measurement_structure",
                "causal_design",
                "identification_report",
                "measurements",
                "panel",
                "validation_report",
                "statistical_model_spec",
                "compiled_ssm",
                "posterior",
                "baseline_report",
                "saved_scenarios"
              ],
              "type": "string",
              "title": "Artifact Id"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/episodes/{workspace_id}/traces": {
      "get": {
        "summary": "Get Trace",
        "description": "Resolve an artifact ``llm_trace_ref`` through the configured trace store.",
        "operationId": "get_trace_api_episodes__workspace_id__traces_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "ref",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ref"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMTrace"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/episodes": {
      "post": {
        "summary": "Start Episode",
        "description": "Ensure the episode workflow exists; optionally seed the `question` root.\n\nIdempotent: attaches to an existing episode or starts a fresh one. Passing\n`question` writes the `question` root artifact with `human` provenance.\nUpload raw data at `POST /api/upload` before running the `raw_data`\ntransition. Returns the same shape as\n`GET /api/episodes/{id}`.",
        "operationId": "start_episode_api_episodes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartEpisodeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Start Episode Api Episodes Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/upload": {
      "post": {
        "summary": "Upload File",
        "description": "Stage one raw input file for the raw_data transition.",
        "operationId": "upload_file_api_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_api_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArtifactEnvelope": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "artifact_id": {
            "type": "string",
            "enum": [
              "question",
              "raw_data",
              "latent_structure",
              "measurement_structure",
              "causal_design",
              "identification_report",
              "measurements",
              "panel",
              "validation_report",
              "statistical_model_spec",
              "compiled_ssm",
              "posterior",
              "baseline_report",
              "saved_scenarios"
            ],
            "title": "Artifact Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "meta": {
            "$ref": "#/components/schemas/ArtifactVersionInfo"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "binary_files": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Binary Files"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "workspace_id",
          "artifact_id",
          "version",
          "meta",
          "payload",
          "binary_files"
        ],
        "title": "ArtifactEnvelope"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}